+++ src/components/UploadFile.vue
@@ -18,8 +18,8 @@
 
 
 <script lang="ts">
- import Spinner from './spinner.vue'
- import * as client from '../ts/client'
+ import Spinner from './spinner.vue';
+ import * as client from '../ts/client';
 
  export default {
    components: {