+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 * TODO: Mobile menu scaling.
 * TODO: Clicking on a menu entry in mobile closes the menu.
 * TODO: File size is checked and errors out before uploading.
+* TODO: Implement common uploadfile logic.
 * Moved codefile to pasty-core.
+* Now using pasty-core serializers.
+* Now using paste api v2.
 
 ## 0.2.5
 * Fix for unicode characters breaking the download button (and subsequently the entire page).